refactor(commands): split server command helpers
